Split explog routines from math.cc
