+++ chainerx_cc/chainerx/routines/CMakeLists.txt
@@ -2,6 +2,7 @@ add_library(chainerx_routines STATIC
     binary.cc
     connection.cc
     creation.cc
+    explog.cc
     hyperbolic.cc
     indexing.cc
     linalg.cc
@@ -22,6 +23,7 @@ install(FILES
     binary.h
     connection.h
     creation.h
+    explog.h
     hyperbolic.h
     indexing.h
     linalg.h
