+++ chainerx_cc/chainerx/routines/math.cc
@@ -19,6 +19,7 @@
 #include "chainerx/kernels/math.h"
 #include "chainerx/macro.h"
 #include "chainerx/routines/creation.h"
+#include "chainerx/routines/explog.h"
 #include "chainerx/routines/logic.h"
 #include "chainerx/routines/manipulation.h"
 #include "chainerx/routines/misc.h"
@@ -423,132 +424,6 @@ Array Divide(const Array& x1, Scalar x2) { return TrueDivide(x1, x2); }
 Array Divide(Scalar x1, const Array& x2) { return TrueDivide(x1, x2); }
 
 Array Reciprocal(const Array& x) { return Scalar{1, GetKind(x.dtype())} / x; }
-
-Array Exp(const Array& x) {
-    Dtype dtype = internal::GetMathResultDtype(x.dtype());
-    Array out = Empty(x.shape(), dtype, x.device());
-
-    {
-        NoBackpropModeScope scope{};
-        x.device().backend().CallKernel<ExpKernel>(x, out);
-    }
-
-    BackwardBuilder bb{"exp", x, out};
-    if (BackwardBuilder::Target bt = bb.CreateTarget(0)) {
-        bt.Define([out_tok = bb.RetainOutput(0)](BackwardContext& bctx) {
-            const Array& out = bctx.GetRetainedOutput(out_tok);
-            bctx.input_grad() = *bctx.output_grad() * out;
-        });
-    }
-    bb.Finalize();
-
-    return out;
-}
-
-Array Expm1(const Array& x) {
-    Dtype dtype = internal::GetMathResultDtype(x.dtype());
-    Array out = Empty(x.shape(), dtype, x.device());
-
-    {
-        NoBackpropModeScope scope{};
-        x.device().backend().CallKernel<Expm1Kernel>(x, out);
-    }
-
-    BackwardBuilder bb{"expm1", x, out};
-    if (BackwardBuilder::Target bt = bb.CreateTarget(0)) {
-        bt.Define([out_tok = bb.RetainOutput(0)](BackwardContext& bctx) {
-            const Array& out = bctx.GetRetainedOutput(out_tok);
-            bctx.input_grad() = *bctx.output_grad() * (out + 1);
-        });
-    }
-    bb.Finalize();
-
-    return out;
-}
-
-Array Exp2(const Array& x) {
-    Dtype dtype = internal::GetMathResultDtype(x.dtype());
-    Array out = Empty(x.shape(), dtype, x.device());
-
-    {
-        NoBackpropModeScope scope{};
-        x.device().backend().CallKernel<Exp2Kernel>(x, out);
-    }
-
-    BackwardBuilder bb{"exp2", x, out};
-    if (BackwardBuilder::Target bt = bb.CreateTarget(0)) {
-        bt.Define([out_tok = bb.RetainOutput(0)](BackwardContext& bctx) {
-            const Array& out = bctx.GetRetainedOutput(out_tok);
-            bctx.input_grad() = *bctx.output_grad() * out * std::log(2.0);
-        });
-    }
-    bb.Finalize();
-
-    return out;
-}
-
-Array Log(const Array& x) {
-    Dtype dtype = internal::GetMathResultDtype(x.dtype());
-    Array out = Empty(x.shape(), dtype, x.device());
-
-    {
-        NoBackpropModeScope scope{};
-        x.device().backend().CallKernel<LogKernel>(x, out);
-    }
-
-    BackwardBuilder bb{"log", x, out};
-    if (BackwardBuilder::Target bt = bb.CreateTarget(0)) {
-        bt.Define([x_tok = bb.RetainInput(0)](BackwardContext& bctx) {
-            const Array& x = bctx.GetRetainedInput(x_tok);
-            bctx.input_grad() = *bctx.output_grad() / x;
-        });
-    }
-    bb.Finalize();
-
-    return out;
-}
-
-Array Log10(const Array& x) {
-    Dtype dtype = internal::GetMathResultDtype(x.dtype());
-    Array out = Empty(x.shape(), dtype, x.device());
-
-    {
-        NoBackpropModeScope scope{};
-        x.device().backend().CallKernel<Log10Kernel>(x, out);
-    }
-
-    BackwardBuilder bb{"log10", x, out};
-    if (BackwardBuilder::Target bt = bb.CreateTarget(0)) {
-        bt.Define([x_tok = bb.RetainInput(0)](BackwardContext& bctx) {
-            const Array& x = bctx.GetRetainedInput(x_tok);
-            bctx.input_grad() = *bctx.output_grad() / x * (1.0 / std::log(10));
-        });
-    }
-    bb.Finalize();
-
-    return out;
-}
-
-Array Log1p(const Array& x) {
-    Dtype dtype = internal::GetMathResultDtype(x.dtype());
-    Array out = Empty(x.shape(), dtype, x.device());
-
-    {
-        NoBackpropModeScope scope{};
-        x.device().backend().CallKernel<Log1pKernel>(x, out);
-    }
-
-    BackwardBuilder bb{"log1p", x, out};
-    if (BackwardBuilder::Target bt = bb.CreateTarget(0)) {
-        bt.Define([x_tok = bb.RetainInput(0)](BackwardContext& bctx) {
-            const Array& x = bctx.GetRetainedInput(x_tok);
-            bctx.input_grad() = *bctx.output_grad() / (1.0 + x);
-        });
-    }
-    bb.Finalize();
-
-    return out;
-}
 
 Array Sigmoid(const Array& x) {
     Dtype dtype = internal::GetMathResultDtype(x.dtype());
