+++ chainerx_cc/chainerx/routines/math.h
@@ -71,13 +71,6 @@ Array TrueDivide(Scalar x1, const Array& x2);
 
 Array Reciprocal(const Array& x);
 
-Array Exp(const Array& x);
-Array Expm1(const Array& x);
-Array Exp2(const Array& x);
-Array Log(const Array& x);
-Array Log10(const Array& x);
-Array Log1p(const Array& x);
-
 Array Sigmoid(const Array& x);
 
 Array Relu(const Array& x);
